Conflicts:
	Web/Areas/Procurement/Controllers/BidController.cs
	Web/Areas/Procurement/Views/Bid/Index.cshtml

+++ Migrations/Cats.DatabaseSchema/Procurement/Tables/BidDetail.sql
@@ -4,8 +4,12 @@
     [RegionID]               INT             NOT NULL,
     [AmountForReliefProgram] DECIMAL (19, 5) NULL,
     [AmountForPSNPProgram]   DECIMAL (19, 5) NULL,
-    [BidDocumentPrice]       MONEY           NOT NULL,
-    [CPO]                    MONEY           NOT NULL,
+
+    [BidDocumentPrice]       REAL            NOT NULL,
+    [CPO]                    REAL            NOT NULL,
+
+    
+
     CONSTRAINT [PK_BidDetail] PRIMARY KEY CLUSTERED ([BidDetailID] ASC),
     CONSTRAINT [FK_BidDetail_AdminUnit] FOREIGN KEY ([RegionID]) REFERENCES [dbo].[AdminUnit] ([AdminUnitID]),
     CONSTRAINT [FK_BidDetail_Bid] FOREIGN KEY ([BidID]) REFERENCES [dbo].[Bid] ([BidID])
